ProposalLLM-main/requirements_converter.txt
python-docx==0.8.11
Pillow==10.0.0
requests==2.31.0
beautifulsoup4==4.12.2
html2image==2.0.1
lxml==4.9.3 